+++ src/components/Sidebar.jsx
@@ -2,10 +2,12 @@ import React from 'react'
 import Link from 'next/link'
 import { BsFillPeopleFill, BsGithub, BsTelegram, BsTwitter, BsDiscord } from 'react-icons/bs'
 import { AiFillHome } from 'react-icons/ai'
-import { BiServer } from 'react-icons/bi'
+import {RiGovernmentFill} from 'react-icons/ri'
+import { BiServer ,BiCoin,BiGroup} from 'react-icons/bi'
 import { MdEditRoad } from 'react-icons/md'
 import { useSelector, useDispatch } from 'react-redux'
 import { updateSidebar } from '@/redux/slices/sidebarSlice'
+import {GiRegeneration} from "react-icons/gi";
 function Sidebar() {
     const dispatch = useDispatch()
     const sidebar = useSelector(state => state.sidebar.value)
@@ -17,11 +19,14 @@ function Sidebar() {
                     <span className="ml-4 text-2xl font-bold">STABILITY</span>
                 </div>
             </Link>
-            <ul className="mt-3">
-                <li><Link href="/"><a className="flex items-center py-4 pl-7 text-xl gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><AiFillHome className="mr-2" />Home</a></Link></li>
-                <li><Link href="/roadmap"><a className="flex items-center py-4 pl-7 text-xl gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><MdEditRoad className="mr-2" />Roadmap</a></Link></li>
-                <li><Link href="/ecosystem"><a className="flex items-center py-4 pl-7 text-xl gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><BsFillPeopleFill className="mr-2" />Ecosystem</a></Link></li>
-                <li><Link href="/pools"><a className="flex items-center py-4 pl-7 text-xl gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><BiServer className="mr-2" />Pools</a></Link></li>
+            <ul className="mt-3" className="overflow-y-auto" style={{height:"calc(100vh - 160px)"}}>
+                <li><Link href="/"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><AiFillHome className="mr-2" />Home</a></Link></li>
+                <li><Link href="/roadmap"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><MdEditRoad className="mr-2" />Roadmap</a></Link></li>
+                <li><Link href="/pools"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><BiServer className="mr-2" />Pools</a></Link></li>
+                <li><Link href="/tokens"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><BiCoin className="mr-2" />Tokens</a></Link></li>
+                <li><Link href="/governance"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><RiGovernmentFill className="mr-2" />Governance</a></Link></li>
+                <li><Link href="/team"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><BiGroup className="mr-2" />Team</a></Link></li>
+                <li><Link href="/generation"><a className="flex items-center py-4 text-xl pl-7 gap-x-2 " onClick={() => { dispatch(updateSidebar(false)) }} ><GiRegeneration className="mr-2" />Generation</a></Link></li>
             </ul>
             <div className="absolute flex flex-col items-center w-72 md:w-56 xl:w-60 bottom-2 gap-y-2">
                 <ul className="flex justify-center text-center gap-x-5">
@@ -30,7 +35,7 @@ function Sidebar() {
                     <li><a href="https://t.me/stabilitydao" target="_blank" rel="noopener noreferrer" ><BsTelegram className="text-3xl " /></a></li>
                     <li><a href="https://discord.gg/R3nnetWzC9" target="_blank" rel="noopener noreferrer" ><BsDiscord className="text-3xl " /></a></li>
                 </ul>
-                <div className="text-sm my-2">
+                <div className="my-1 text-sm">
                     © 2021 Stability
                 </div>
             </div>
