+++ src/components/view/Governance.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+
+function Governance() {
+    return (
+        <section className="dark:bg-black dark:text-white h-calc">
+            <div className="container p-4">
+                <h1 className="mb-4 text-4xl font-semibold leading-10 tracking-wide text-center text-indigo-500 sm:text-6xl font-Roboto">Governance</h1>
+                <div className="max-w-lg mx-auto">
+                    <div>
+                        <article className="mb-8">
+                            <a className="py-4 overflow-hidden rounded-xl lg:p-6 ">
+                                <p className="text-lg">
+                                    PROFIT token holders with sufficient voting power can directly participate in the governance of the Stability protocol. These will be the only investors who can collectively change the organization’s resource distribution (mentioned below) and participate in critical decision-making processes involving the direction of the entire ecosystem.
+                                </p>
+                            </a>
+                        </article>
+                    </div>
+                </div>
+            </div>
+        </section>
+    )
+}
+
+export default Governance
