--- src/pages/ecosystem.js
@@ -1,6 +1,0 @@
-import {Ecosystem } from '../components/view'
-function ecosystem() {
-    return <Ecosystem/>
-}
-
-export default ecosystem
