+++ src/pages/generation.js
@@ -1,0 +1,7 @@
+import { Generation } from '../components/view'
+
+function generation() {
+    return <Generation />
+}
+
+export default generation
