+++ src/pages/governance.js
@@ -1,0 +1,7 @@
+import { Governance } from '../components/view'
+
+function governance() {
+    return <Governance />
+}
+
+export default governance
