+++ src/pages/team.js
@@ -1,0 +1,7 @@
+import { Team } from '../components/view'
+
+function team() {
+    return <Team />
+}
+
+export default team
