+++ src/pages/tokens.js
@@ -1,0 +1,7 @@
+import { Tokens } from '../components/view'
+
+function tokens() {
+    return <Tokens />
+}
+
+export default tokens
